Remove circular dependency and move backup ui to its module

+++ feature-scenario-config/src/main/java/com/buzbuz/smartautoclicker/feature/scenario/config/data/base/IdentifierCreator.kt
@@ -22,7 +22,7 @@ import com.buzbuz.smartautoclicker.domain.model.Identifier
 internal class IdentifierCreator {
 
     /** The last generated domain id for an item. */
-    private var lastGeneratedDomainId: Int = 0
+    private var lastGeneratedDomainId: Long = 0
 
     /** */
     fun generateNewIdentifier(): Identifier =
